requirements.txt
binance-connector
pandas_ta
python-dotenv
plotly
telebot
kaleido
scipy
numpy==1.26.4
setuptools
ipython
